project/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.4
django-allauth==0.54.0
django-appconf==1.0.5
django-bootstrap5==23.3
django-filter==23.2
django-tinymce==3.6.1
django-user-accounts==3.2.0
idna==3.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
